renamed ping to stats and added more informations
